+++ lib/widgets/Categories/cone_category.dart
@@ -1,13 +1,15 @@
-import 'package:final_project/screens/cartpage.dart';
+import 'package:final_project/screens/cart_page.dart';
 import 'package:final_project/screens/detail_page.dart';
 import 'package:final_project/screens/home_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_rating_bar/flutter_rating_bar.dart';
 
-import 'categoryitems.dart';
+import '../../model/category_items.dart';
 
-class SpecialPage extends StatelessWidget {
-  const SpecialPage({super.key});
+class ConesPage extends StatelessWidget {
+  const ConesPage({
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -34,7 +36,7 @@ class SpecialPage extends StatelessWidget {
             },
           ),
           title: const Text(
-            "Specials",
+            "Cones",
             style: TextStyle(fontWeight: FontWeight.bold, fontSize: 30),
           ),
           centerTitle: true,
@@ -42,9 +44,9 @@ class SpecialPage extends StatelessWidget {
             decoration: BoxDecoration(
               gradient: LinearGradient(
                 colors: [
-                  Colors.pink.shade100,
-                  Colors.purple.shade300,
-                ],
+                  Colors.teal.shade200,
+                  Colors.green
+                ], // Add your desired gradient colors
                 begin: Alignment.topCenter,
                 end: Alignment.bottomCenter,
               ),
@@ -59,9 +61,9 @@ class SpecialPage extends StatelessWidget {
             crossAxisSpacing: 20.0,
             childAspectRatio: screenWidth > 600 ? 0.7 : 0.8,
           ),
-          itemCount: homes.length > 13 ? 4 : homes.length,
+          itemCount: homes.length > 5 ? 5 : homes.length,
           itemBuilder: (context, index) {
-            Home home = homes[index + 13];
+            Home home = homes[index];
             return InkWell(
               onTap: () {
                 //yaha se routing krwa kr mein detailpage me jarha hu simple
@@ -119,7 +121,7 @@ class SpecialPage extends StatelessWidget {
                             onTap: () {
                               Navigator.of(context).push(MaterialPageRoute(
                                 builder: (context) => CartPage(
-                                  cartItems: DetailPage.cartitems,
+                                  cartItems: DetailPage.cartItems,
                                 ),
                               ));
                             },
